requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
boto3==1.26.151
botocore==1.29.151
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
distlib==0.3.6
Django==4.2.3
django-ckeditor==6.5.1
django-cors-headers==3.10.0
django-environ==0.4.5
django-filter==22.1
django-js-asset==2.0.0
django-mssql-backend==2.8.1
django-redis==5.2.0
django-simple-history==3.3.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.12.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.0
Markdown==3.4.1
multidict==6.0.4
mutagen==1.46.0
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
twilio==8.2.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
whitenoise==6.4.0
yarl==1.9.2
zipp==3.15.0
